DIALOGUE SYSTEM v0.9
Right now, the player has to press space to get the UI to show up and Q to continue/skip/

Yes, I'll find a way to make the UI show automatically without the space.

In order to use the dialogue, copy the Controller/Canvas/You from the Test scene, and listen to the instructions.
Create the Conversation Node
Create Dialogue Nodes
Win?